Created blueprint extension and changed base script to be able to run as tapermonkey addon and as extension
