+++ conf.yml
@@ -1,0 +1,37 @@
+info:
+  name: "PteroSort Category"
+  identifier: "pterosort"
+  description: "Pterodactyl server sorter with categories"
+  flags: ""
+  version: "1.3.0"
+  target: "beta-2024-12"
+  author: "Ricman"
+  icon: "folder_icon.png"
+  website: "https://github.com/Ricman-MC/PteroSort"
+
+admin:
+  view: "view.blade.php"
+  controller: "controller.php"
+  css: ""
+  wrapper: ""
+
+dashboard:
+  css: ""
+  wrapper: "dashboard/wrapper.blade.php"
+  components: ""
+
+data:
+  directory: ""
+  public: ""
+  console: ""
+
+requests:
+  views: ""
+  app: "controllers"
+  routers:
+    application: ""
+    client: ""
+    web: "routes.php"
+
+database:
+  migrations: "migrations"
