+++ view.blade.php
@@ -1,0 +1,13 @@
+<div class="box box-info">
+  <div class="box-header with-border">
+    <h3 class="box-title">Info</h3>
+  </div>
+  <div class="box-body">
+    <p>
+      <b>{name}</b> extension made by <b>{author}</b>.<br>
+      Build for pterodactyl panel version <b>1.11.10</b>.<br>
+      It enables the function to create categories and ordering for servers within the dashboard.<br>
+      Categories and orders are saved on user basis.
+    </p>
+  </div>
+</div>
